Migrate to TypeScript

+++ src/main/worker.ts
@@ -1,3 +1,5 @@
+import Logger from "./logger";
+import { ChildProcess } from "child_process";
 const { spawn, execSync } = require("child_process");
 const os = require("os");
 const fs = require("fs-extra");
@@ -5,8 +7,15 @@ const path = require("path");
 const { app, ipcMain } = require("electron");
 const log = require("electron-log");
 
-class Worker {
-  constructor(logger) {
+class Controller {
+  status: string;
+  executableDirectory: string;
+  executablePath: string;
+  child: null | ChildProcess;
+  logger: Logger;
+  userDataPath: string;
+  configPath: string;
+  constructor(logger: Logger) {
     this.status = "stopped";
     let executableName = os.platform() === "darwin" ? "v2ray" : "v2ray.exe";
     if (global.ROOT.indexOf("app.asar") > 0) {
@@ -33,7 +42,7 @@ class Worker {
     this.userDataPath = app.getPath("userData");
     this.configPath = path.join(this.userDataPath, "v2ray.json");
     this.initConfig();
-    ipcMain.on("restart", event => {
+    ipcMain.on("restart", () => {
       this.restart();
     });
   }
@@ -42,22 +51,24 @@ class Worker {
     log.info("Starting worker ", this.executablePath);
     log.info("With config", this.configPath);
     this.child = spawn(this.executablePath, ["-config", this.configPath]);
-    this.child.stdout.on('data', data => {
-      this.logger.append(data)
-      })
-    this.child.stderr.on('data', data => {
-      this.logger.append(data)
-    })
+    if (this.child) {
+      this.child.stdout.on("data", data => {
+        this.logger.append(data);
+      });
+      this.child.stderr.on("data", data => {
+        this.logger.append(data);
+      });
+    }
   }
 
   restart() {
     log.info("Restarting worker ", this.executablePath);
-    this.child.kill();
+    this.child && this.child.kill();
     this.start();
   }
 
   stop() {
-    this.child.kill();
+    this.child && this.child.kill();
   }
 
   initConfig() {
@@ -67,7 +78,7 @@ class Worker {
         global.ROOT,
         "assets",
         "v2ray",
-        "v2ray",
+        `v2ray-${os.platform() === "darwin" ? "macos" : "win"}`,
         "config.json.default"
       );
       fs.copySync(defaultConfigPath, this.configPath);
@@ -75,4 +86,4 @@ class Worker {
   }
 }
 
-exports.Worker = Worker;
+export { Controller };
